requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
contourpy==1.3.3
cycler==0.12.1
discord.py==2.6.4
disnake==2.9.2
fonttools==4.62.1
frozenlist==1.4.1
idna==3.8
kiwisolver==1.5.0
matplotlib==3.10.8
multidict==6.1.0
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
yarl==1.11.1